+++ app/Record.php
@@ -59,7 +59,7 @@ class Record extends Model
         'ttl'      => 'integer',
         'type'     => 'string',
         'priority' => 'integer',
-        'data'     => 'string'
+        'data'     => 'string',
     ];
     /**
      * The attributes that should be casted to null if is empty.
@@ -82,7 +82,7 @@ class Record extends Model
      *
      * @return array
      */
-    public static function getAllValidRecordTypes() : array
+    public static function getAllValidRecordTypes(): array
     {
         return [
             'A'     => trans('record/model.types_mapper.A'),
@@ -117,16 +117,6 @@ class Record extends Model
     }
 
     /**
-     * Set the Record's data lowercase.
-     *
-     * @param  string $value
-     */
-    public function setDataAttribute(string $value)
-    {
-        $this->attributes['data'] = strtolower($value);
-    }
-
-    /**
      * A record belongs to a zone.
      *
      * This is a one-to-one relationship.
@@ -142,7 +132,7 @@ class Record extends Model
      * @return string
      * @codeCoverageIgnore
      */
-    public function getResourceRecord() : string
+    public function getResourceRecord(): string
     {
         if ($this->ttl) {
             return sprintf("%-40s %d\tIN\t%s\t%s", $this->name, $this->ttl, $this->type, $this->data);