+++ resources/lang/en/record/model.php
@@ -30,6 +30,7 @@ return [
         'AAAA'  => 'AAAA (IPv6 address)',
         'CNAME' => 'CNAME (canonical name)',
         'MX'    => 'MX (mail exchange)',
+        'NAPTR' => 'NAPTR (name authority pointer)',
         'NS'    => 'NS (name server)',
         'PTR'   => 'PTR (pointer)',
         'SRV'   => 'SRV (service locator)',