+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## Unreleased
 
+## 0.8.0 - 2018-05-20
+### Changed
+-  Improve Dockerfile creation to speed it up
+
+### Fixed
+-  Fix bower dependency problems (#42)
+
 ## 0.7.1 - 2017-08-26
 
 ### Changed